TEMPORARY HACK: Remove dependency from imx_thermal.c to CPU frequency policy.
The imx_thermal.c tries to attach to a CPU frequency policy to
lower CPU frequency on overheating. This policy is not found (unsure
why yet) but neither do we want this now, while we are debugging a
potential heat related problem.

This commit is destined to be reverted in the future.

Signed-off-by: Leon Woestenberg <leon@sidebranch.com>